fix char issues

+++ crates/warp_tui/src/editor_element.rs
@@ -280,7 +280,7 @@ impl TuiEditorElement {
         // The first visible row is model-side scroll state; unwindowed
         // consumers always render from the top.
         let first_visible_row = if self.viewport_rows.is_some() {
-            char_cell.scroll_offset()
+            char_cell.scroll_offset() as usize
         } else {
             0
         };
@@ -295,16 +295,13 @@ impl TuiEditorElement {
             // is part of the layout, so include it when sizing and windowing.
             let cursor = lattice.offset_to_display_point(cursor_offset);
             let total_rows = if self.editable {
-                cursor.map_or(rows.len(), |cursor| rows.len().max(cursor.row as usize + 1))
+                cursor.map_or(rows.len(), |cursor| rows.len().max(cursor.row + 1))
             } else {
                 rows.len()
             };
 
             let (visible_start, visible_rows) = match self.viewport_rows {
-                Some(max_rows) => (
-                    first_visible_row as usize,
-                    (max_rows as usize).min(total_rows),
-                ),
+                Some(max_rows) => (first_visible_row, (max_rows as usize).min(total_rows)),
                 None => (0, total_rows),
             };
             let visible_end = (visible_start + visible_rows).min(total_rows);
@@ -349,9 +346,8 @@ impl TuiEditorElement {
         if let Some(cursor) = cursor {
             self.cursor_col = cursor.col + self.gutter_cols;
             self.cursor_row_in_view = cursor.row.saturating_sub(first_visible_row) as u16;
-            self.cursor_visible = self.editable
-                && cursor.row >= first_visible_row
-                && (cursor.row as usize) < visible_end.max(1);
+            self.cursor_visible =
+                self.editable && cursor.row >= first_visible_row && cursor.row < visible_end.max(1);
         } else {
             self.cursor_col = 0;
             self.cursor_row_in_view = 0;
